trains: inputs: add InteractiveInput class
